requirements.txt
flask
flask-cors
openai
mtranslate
python-dotenv
nltk
empath